+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](
 ### Added
 
 - Defined tokens and transactions table and draft security rules.
-- List venues using a currency
+- List venues using a currency.
+- List minters on a currency.
 
 ### Maintenance